refactor: inline hash digest string rules
